Projects/Deploying-a-Machine-Learning-Model-on-Heroku-with-FastAPI/requirements.txt
numpy
pandas
scikit-learn==1.1.3
pytest
requests
fastapi==0.63.0
uvicorn
gunicorn
PyYAML
flake8
httpx
dvc
dvc-s3
